# miscellaneous/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Evently Booking Platform API",
...
    "contact": {
      "name": "Evently Support",
      "email": "droplakshya@gmail.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/v1/events/": {
      "get": {
        "tags": ["events"],
        "summary": "List Events",
        "description": "Get list of events with filtering and pagination.\n\nThis endpoint supports various filtering options:\n- Search by name, description, or venue\n- Filter by venue, date range, price range\n- Show only available or active events",
        "operationId": "list_events_api_v1_events__get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size",
              "default": 20,
              "title": "Size"
            },
            "description": "Page size"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Search in event name, description, or venue",
              "title": "Search"
            },
            "description": "Search in event name, description, or venue"
          },
          {
            "name": "venue",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by venue",
              "title": "Venue"
            },
            "description": "Filter by venue"
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter events from this date (ISO format)",
              "title": "Date From"
            },
            "description": "Filter events from this date (ISO format)"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter events until this date (ISO format)",
              "title": "Date To"
            },
            "description": "Filter events until this date (ISO format)"
          },
          {
            "name": "min_price",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0,
              "description": "Minimum ticket price",
              "title": "Min Price"
            },
            "description": "Minimum ticket price"
          },
          {
            "name": "max_price",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0,
              "description": "Maximum ticket price",
              "title": "Max Price"
            },
            "description": "Maximum ticket price"
          },
          {
            "name": "available_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Show only events with available capacity",
              "default": true,
              "title": "Available Only"
            },
            "description": "Show only events with available capacity"
          },
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Show only active events",
              "default": true,
              "title": "Active Only"
            },
            "description": "Show only active events"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["events"],
        "summary": "Create Event",
        "description": "Create a new event.\n\nOnly admin users can create events.",
        "operationId": "create_event_api_v1_events__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/{event_id}": {
      "get": {
        "tags": ["events"],
        "summary": "Get Event",
        "description": "Get event details by ID.",
        "operationId": "get_event_api_v1_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["events"],
        "summary": "Update Event",
        "description": "Update an existing event.\n\nOnly admin users can update events.",
        "operationId": "update_event_api_v1_events__event_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["events"],
        "summary": "Delete Event",
        "description": "Delete an event.\n\nOnly admin users can delete events.\nEvents with confirmed bookings cannot be deleted.",
        "operationId": "delete_event_api_v1_events__event_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/search/{search_term}": {
      "get": {
        "tags": ["events"],
        "summary": "Search Events",
        "description": "Search events by name, description, or venue.",
        "operationId": "search_events_api_v1_events_search__search_term__get",
        "parameters": [
          {
            "name": "search_term",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Search Term"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Maximum number of results",
              "default": 10,
              "title": "Limit"
            },
            "description": "Maximum number of results"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventResponse"
                  },
                  "title": "Response Search Events Api V1 Events Search  Search Term  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/popular/list": {
      "get": {
        "tags": ["events"],
        "summary": "Get Popular Events",
        "description": "Get popular events based on booking count.",
        "operationId": "get_popular_events_api_v1_events_popular_list_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Maximum number of results",
              "default": 10,
              "title": "Limit"
            },
            "description": "Maximum number of results"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventResponse"
                  },
                  "title": "Response Get Popular Events Api V1 Events Popular List Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/upcoming/list": {
      "get": {
        "tags": ["events"],
        "summary": "Get Upcoming Events",
        "description": "Get upcoming events with available capacity.",
        "operationId": "get_upcoming_events_api_v1_events_upcoming_list_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Maximum number of results",
              "default": 10,
              "title": "Limit"
            },
            "description": "Maximum number of results"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventResponse"
                  },
                  "title": "Response Get Upcoming Events Api V1 Events Upcoming List Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/waitlist/admin/event/{event_id}": {
      "get": {
        "tags": ["waitlist"],
        "summary": "Get Event Waitlist",
        "description": "Get waitlist entries for a specific event (admin only).\n\n- **event_id**: ID of the event\n- **status**: Optional filter by waitlist status\n- **limit**: Maximum number of entries to return (1-500)\n- **offset**: Number of entries to skip for pagination",
        "operationId": "get_event_waitlist_api_v1_waitlist_admin_event__event_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WaitlistStatus"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by waitlist status",
              "title": "Status"
            },
            "description": "Filter by waitlist status"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Maximum number of entries to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of entries to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of entries to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of entries to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WaitlistResponse"
                  },
                  "title": "Response Get Event Waitlist Api V1 Waitlist Admin Event  Event Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "Event name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Event description"
          },
          "venue": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Venue",
            "description": "Event venue"
          },
          "event_date": {
            "type": "string",
            "format": "date-time",
            "title": "Event Date",
            "description": "Event date and time"
          },
          "total_capacity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Total Capacity",
            "description": "Total event capacity"
          },
          "price": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string"
              }
            ],
            "title": "Price",
            "description": "Event ticket price"
          },
          "has_seat_selection": {
            "type": "boolean",
            "title": "Has Seat Selection",
            "description": "Whether event has seat selection",
            "default": false
          }
        },
        "type": "object",
        "required": ["name", "venue", "event_date", "total_capacity", "price"],
        "title": "EventCreate",
        "description": "Schema for creating a new event."
      },
      "EventListResponse": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventResponse"
            },
            "type": "array",
            "title": "Events"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": ["events", "total", "page", "size", "pages"],
        "title": "EventListResponse",
        "description": "Schema for paginated event list response."
      },
...
      "EventResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "Event name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Event description"
          },
          "venue": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Venue",
            "description": "Event venue"
          },
          "event_date": {
            "type": "string",
            "format": "date-time",
            "title": "Event Date",
            "description": "Event date and time"
          },
          "total_capacity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Total Capacity",
            "description": "Total event capacity"
          },
          "price": {
            "type": "string",
            "title": "Price",
            "description": "Event ticket price"
          },
          "has_seat_selection": {
            "type": "boolean",
            "title": "Has Seat Selection",
            "description": "Whether event has seat selection",
            "default": false
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "available_capacity": {
            "type": "integer",
            "title": "Available Capacity"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "is_sold_out": {
            "type": "boolean",
            "title": "Is Sold Out"
          },
          "capacity_utilization": {
            "type": "number",
            "title": "Capacity Utilization"
          }
        },
        "type": "object",
        "required": [
          "name",
          "venue",
          "event_date",
          "total_capacity",
          "price",
          "id",
          "available_capacity",
          "is_active",
          "version",
          "created_at",
          "updated_at",
          "is_sold_out",
          "capacity_utilization"
        ],
        "title": "EventResponse",
        "description": "Schema for event response."
      },
      "EventUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "venue": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue"
          },
          "event_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Date"
          },
          "total_capacity": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Capacity"
          },
          "price": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "has_seat_selection": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has Seat Selection"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "EventUpdate",
        "description": "Schema for updating an existing event."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "WaitlistStatus": {
        "type": "string",
        "enum": ["active", "notified", "expired", "converted"],
        "title": "WaitlistStatus",
        "description": "Enumeration for waitlist status."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}